refactor: rename construct-constructor-signature to construct-constructor-property and allow additional parameters
