+++ docs/rules/index.md
@@ -148,16 +148,6 @@ Currently we support the following rules:
   </li>
   <li class="rule-item">
     <div class="rule-info">
-      <a href="/rules/construct-constructor-signature" class="rule-name">construct-constructor-signature</a>
-      <span class="rule-description">Enforces that constructors of classes extending Construct have the signature 'scope, id' or 'scope, id, props'</span>
-    </div>
-    <div class="rule-status">
-      <span class="status-icon recommended">✅</span>
-      <span class="status-icon fixable"/>
-    </div>
-  </li>
-  <li class="rule-item">
-    <div class="rule-info">
       <a href="/rules/no-variable-construct-id" class="rule-name">no-variable-construct-id</a>
       <span class="rule-description">Disallow variables in Construct IDs</span>
     </div>
@@ -224,6 +214,16 @@ Currently we support the following rules:
     <div class="rule-status">
       <span class="status-icon recommended">✅</span>
       <span class="status-icon fixable">🔧</span>
+    </div>
+  </li>
+  <li class="rule-item">
+    <div class="rule-info">
+      <a href="/rules/construct-constructor-property" class="rule-name">construct-constructor-property</a>
+      <span class="rule-description">Enforces that constructors of classes extending Construct have the property names 'scope, id' or 'scope, id, props'</span>
+    </div>
+    <div class="rule-status">
+      <span class="status-icon recommended">✅</span>
+      <span class="status-icon fixable"/>
     </div>
   </li>
   <li class="rule-item">
