+++ src/index.ts
@@ -1,4 +1,4 @@
-import { constructConstructorSignature } from "./rules/construct-constructor-signature";
+import { constructConstructorProperty } from "./rules/construct-constructor-property";
 import { noClassInInterface } from "./rules/no-class-in-interface";
 import { noConstructStackSuffix } from "./rules/no-construct-stack-suffix";
 import { noImportPrivate } from "./rules/no-import-private";
@@ -14,7 +14,7 @@ import { requirePassingThis } from "./rules/require-passing-this";
 import { requirePropsDefaultDoc } from "./rules/require-props-default-doc";
 
 const rules = {
-  "construct-constructor-signature": constructConstructorSignature,
+  "construct-constructor-property": constructConstructorProperty,
   "no-class-in-interface": noClassInInterface,
   "no-construct-stack-suffix": noConstructStackSuffix,
   "no-parent-name-construct-id-match": noParentNameConstructIdMatch,
@@ -34,11 +34,11 @@ const configs = {
   recommended: {
     plugins: ["cdk"],
     rules: {
-      "cdk/construct-constructor-signature": "error",
+      "cdk/construct-constructor-property": "error",
       "cdk/no-class-in-interface": "error",
       "cdk/no-construct-stack-suffix": [
         "error",
-        { disallowedSuffixes: ["Construct"] },
+        { disallowedSuffixes: ["Stack"] },
       ],
       "cdk/no-parent-name-construct-id-match": "error",
       "cdk/no-public-class-fields": "error",
@@ -55,7 +55,7 @@ const configs = {
   strict: {
     plugins: ["cdk"],
     rules: {
-      "cdk/construct-constructor-signature": "error",
+      "cdk/construct-constructor-property": "error",
       "cdk/no-class-in-interface": "error",
       "cdk/no-construct-stack-suffix": "error",
       "cdk/no-parent-name-construct-id-match": "error",
